location helper function:
